Add fetchAllCompanies feature and fix error
There was a problem in the FindAll() function inside companyRepo that was caused by running the findByID() function inside the rows.next block of the 'select all ids' query. I extracted this part out and added a slice of all company ids to make it work. Altough my first solution might work if cursors are used, not sure though